+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed an issue with the `Input` component where the `outline-on-accent` variant would not have the correct colors.
+- Fixed an issue with the `Link` component where the `menu` variant would not have the correct colors.
+- Fixed an issue with the `Popover`, `Table` and `RadioCard` component not having the correct `background-color`.
+
 ## [0.3.0-beta.0] - 2023-06-06
 
 ### Changed